Mouvements en documents, reprise depuis le classeur, paquets clients
LE MOUVEMENT DEVIENT UN DOCUMENT. Il se lisait ligne a ligne, comme un grand
livre comptable : la vue de l'auditeur, pas celle du magasin, qui manipule des
camions et des equipes. Devant le livre seul, il fallait additionner de tete les
lignes d'un meme bon pour savoir combien de kilos etaient sortis.

  * `GET /api/mouvements/documents` rend un bon par ligne, avec ses totaux ;
    `GET /api/mouvements/{id}` rend le dossier complet, en un appel, parce que
    c'est ce qu'imprime un bon d'entree ou de sortie. Le grand livre reste servi
    par `/api/mouvements` : les deux lectures sont legitimes.
  * `mouvement.responsable` dit QUI A REMIS OU RECU, par opposition a
    `id_utilisateur` qui dit qui a TAPE. Le magasinier saisit souvent pour un
    chef d'equipe ou un chauffeur, et c'est ce dernier qu'on cherche quand un
    ecart apparait trois jours plus tard.
  * `date_mouvement` devient saisissable : un mouvement du samedi saisi le lundi
    porte ses deux dates, et l'ecart est imprime. Le cacher serait un mensonge
    par omission. C06 refuse toujours une date future.
  * `nb_bobines` / `nb_palettes` portent les colis REELLEMENT COMPTES. La
    quantite en palettes se convertit en kg par un facteur theorique ; une
    palette incomplete reste une palette a manutentionner.
  * Le rebut (motif R4) est totalise a part : additionner l'excedent
    reutilisable et la chute perdue ferait croire que tout revient en stock.

LE REFERENTIEL VIENT DU CLASSEUR, PLUS D'UN FICHIER FIGE. `seed_100` etait une
photographie ; « GESTION FIL.xlsx » est la source vivante. `charger.py
--production` charge desormais les comptes et l'identite, et `gestionfil-import`
apporte catalogue, fournisseurs, qualites et recettes. Charger les deux ferait
entrer en concurrence deux verites sur les memes references.

PAQUETS CLIENTS. Le serveur sert les installateurs qu'on depose dans
`/opt/gestionfil/telechargements` — le nom du fichier est le catalogue. Chaque
telechargement est inscrit : c'est la reponse a « sur quelle version tourne ce
poste », la seule question qu'on se pose apres coup.

L'IDENTITE DE L'ENTREPRISE SORT DE LA BASE. La barre de pied porte adresse et
telephone, qu'on cherchait sans jamais les trouver ; les etats imprimes portent
RC, IF, patente et CNSS — un document commercial marocain qui en manque n'a pas
de valeur.

TROIS DEFAUTS TROUVES EN CHEMIN, CORRIGES ICI :

  * le service ecrivait l'URL de la base AVEC SON MOT DE PASSE dans le journal
    systemd, a chaque demarrage (`url_sans_mot_de_passe`) ;
  * `unites::charger` lisait des colonnes `numeric` vers `f64` sans cast :
    toute saisie en palettes tombait en « erreur interne » ;
  * `publier.py` emportait `target/` a chaque envoi, donc chaque publication
    payait une compilation complete de quatre minutes.

Ecran de connexion : fond « fils de lumiere » a la place du motif de tapis, et
disposition en colonne sous 64rem — la carte sortait de l'ecran sur telephone et
tablette. Verifie par capture aux quatre largeurs.

+++ backend/src/domain/inventaire.rs
@@ -128,7 +128,7 @@ pub async fn cloturer(
         "SELECT COUNT(*) FROM ligne_inventaire
           WHERE id_inventaire = $1
             AND ecart_pct IS NOT NULL
-            AND abs(ecart_pct) > (SELECT CAST(valeur_courante AS REAL) FROM parametre
+            AND abs(ecart_pct) > (SELECT CAST(valeur_courante AS numeric) FROM parametre
                                    WHERE code_parametre = 'P_TolerEcartPesee')
             AND (motif_ecart IS NULL OR trim(motif_ecart) = '')",
     )
